Error and message handling was implemented
