+++ Online Learning Management/Domain/Entities/Students.cs
@@ -2,15 +2,10 @@
 {
     public class Students
     {
-  
-    
         public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
-        public int age { get; set; }
+        public int Age { get; set; }
         public int Score { get; set; } 
-        
-    
-        
     }
 }
