+++ Online Learning Management/Infrastructure/Data/ApplicationDbContext.cs
@@ -1,5 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
+
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Online_Learning_Management.Domain.Entities;
+
+using Online_Learning_Management.Infrastructure.Repositories.Student;
 
 namespace Online_Learning_Management.Infrastructure.Data
 {
@@ -11,5 +15,11 @@ namespace Online_Learning_Management.Infrastructure.Data
         public DbSet<User> Users { get; set; }
 
         public DbSet<Students> Students { get; set; }
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            base.OnModelCreating(modelBuilder);
+            modelBuilder.ApplyConfiguration(new StudentsConfiguration());
+            // Puedes agregar más configuraciones aquí si tienes más entidades
+        }
     }
 }
