+++ appsettings.json
@@ -1,0 +1,13 @@
+
+{
+"ConnectionStrings": {
+    "DefaultConnection": "Server=tcp:learningmanagement.database.windows.net,1433;Initial Catalog=learning_management_db;Persist Security Info=False;User ID=nicoZela23;Password=Learning_management_db1;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*"
+}
